requirements.txt
slack-sdk>=3.27.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
openai>=1.3.0
groq>=0.4.0
pydantic>=2.0.0
requests>=2.31.0
apscheduler>=3.10.0

